Address issues from PR feedback
