+++ tests/test_storages.py
@@ -1,14 +1,35 @@
 import pytest
 from django.core.files.base import ContentFile
 
-from s3file import storages
+from s3file.storages_optimized import S3OptimizedUploadStorage
+
+
+class S3OptimizedMockStorage(S3OptimizedUploadStorage):
+    created_objects = {}
+
+    def _compress_content(self, content):
+        return content
+
+    class bucket:
+        name = "test-bucket"
+
+        class Object:
+            def __init__(self, key):
+                self.key = key
+                self.copy_from_bucket = None
+                self.copy_from_key = None
+                S3OptimizedMockStorage.created_objects[self.key] = self
+
+            def copy(self, s3_object, ExtraArgs):
+                self.copy_from_bucket = s3_object["Bucket"]
+                self.copy_from_key = s3_object["Key"]
 
 
 class TestStorages:
     url = "/__s3_mock__/"
 
     def test_post__save_optimized(self):
-        storage = storages.S3OptimizedMockStorage()
+        storage = S3OptimizedMockStorage()
         obj = storage.bucket.Object("tmp/s3file/s3_file.txt")
 
         class Content:
@@ -26,7 +47,7 @@ class TestStorages:
         assert stored_object.copy_from_key == "tmp/s3file/s3_file.txt"
 
     def test_post__save_optimized_gzip(self):
-        storage = storages.S3OptimizedMockStorage()
+        storage = S3OptimizedMockStorage()
         obj = storage.bucket.Object("tmp/s3file/s3_file.css")
         storage.gzip = True
 
@@ -45,9 +66,9 @@ class TestStorages:
         assert stored_object.copy_from_key == "tmp/s3file/s3_file.css"
 
     def test_post__save_optimized_fail(self):
-        storage = storages.S3OptimizedMockStorage()
+        storage = S3OptimizedMockStorage()
 
-        with pytest.raises(RuntimeError) as excinfo:
+        with pytest.raises(TypeError) as excinfo:
             storage._save("tmp/s3file/s3_file_copied.txt", ContentFile(b"s3file"))
 
         assert "The content object must be a S3 object and contain a valid key." in str(
